eval/requirements.txt
ragas
langchain-openai
datasets
pandas
requests
python-dotenv
